Checkpoint.  Beginning of refactoring into a callback-based architecture, to support an event-driven I/O core.
WARNING: At this point in the conversion, some of the programs will
not run indefinitely, because the partial conversion keeps the call
stack from ever unwinding all the way.  I'm willing to tolerate this
temporarily as this allows me to keep running regression tests during
the conversion process, but it does mean that as of this checkin the
code is not even remotely suitable for non-testing use until the I/O
core rewrite is finished.

svn path=/rpkid/rpki/https.py; revision=2345