Trying to guess what we might have installed after the fact is too flakey, given all the whacky "custom layout" hacks added to distutils by Debian-derived releases and the complete absence of /usr/local from the Python search path on Fedora.  What a can of worms.
So instead we just ask distutils to record the list of files it
installed, and deinstallation consists of removing those files.  Not
ideal, but hard to do better without a lot of work, given the
constraints, and probably not worth trying, since in the long run
normal users ought to be using platform-specific packaging systems
(apt, yum, portupgrade, macports, ...) to install and deinstall
anyway.

svn path=/rpkid/Makefile.in; revision=3636