Fixes #93.  New config parameter allow-digest-mismatch, default is warn and allow.  Digest mismatch is now a warning message only.
Missing object that's listed in manifest is now logged as a serious
problem.  At the moment, we do this for missing objects in both
current and backup generations.  I'm not sure that's right, might make
more sense to log this only for current generation, but let's try
running this way for a little while to see what patterns show up.

svn path=/rcynic/README; revision=4007