Major cleanup of internal databases:
rc->rsync_cache and rc->dead_host_cache have merged, acquired
timestamps, and been renamed rc->rsync_history.  Revised code is
simpler, even.  This is most of #110, but addition of connection data
to XML output still needs to be done.

rc->backup_cache is gone, as the information it held was a subset of
what we were already keeping in rc->validation_status.  This was most
of the groundwork for #109, but code still needs tweaking to perform
this sort of check for objects other than certificates.

svn path=/trunk/; revision=4058