Clean up exceptions in rpki.x509.CMS_object.verify().
Up-down specification explictly allows certificates and CRLs we
consider extraneous to be included in CMS messages.  I think that any
verification model which relies on these is flawed, so I don't use
them, in fact, nobody does, but I was unable to convince my co-authors
to change the specification.  Sigh.  So the up-down code has to allow
these.  In theory, OpenSSL's verification code should have no problem
picking the useful bits out of the garbage, since that's a normal part
of OpenSSL's job when acting as a TLS client, so we just pass the mess
along to OpenSSL and let OpenSSL sort it out.

This closes #9.

svn path=/trunk/; revision=4287