Seems that the earlier istallation success with new Apache configuration on Precise was a false positive caused by an orphaned symlink.  Rewrite again to separate Debian and Ubuntu so we can do version number checks in a halfway sane fashion.
Most likely the real underlying issue is Debian-family packaging for
Apache 2.2 vs Apache 2.4, so there may be possible to simplify this
again at some point, but in theory this will do for now.

Need to re-test on Trusty and Wheezy to make sure the fix for Precise
didn't break them.  See #720.

svn path=/trunk/; revision=5995