Task system now working with Tornado.  Two new problems: some kind of UTF-8 whining on what are supposed to be binary fields that's probably the result of a MySQL upgrade, and CMS Replay exceptions due to the pseudo-random order in which HTTP client connections run in Tornado.
The UTF-8 mess is probably a good reason to change over to Django's
native binary field type, which we were going to want to do anyway.

The CMS Replay problem is not Tornado's fault: we probably would have
seen it in the old code were it not for an accidental side effect of a
long-since-abandoned attempt to use persistent HTTP connections.  The
fix is probably to serialize requests to a particular host using use a
tornaodo.queue.Queue() object, or something like that.

svn path=/branches/tk705/; revision=6143