Tweak binary field types to get SQLite3 working as a Django ORM backend.  Switch yamltest's default database configuration to sqlite3.
MySQL still has character set issues, which are almost certainly to do
with the communication channel rather than the database tables.  It's
possible that one of the newer DB API drivers for MySQL fixes this,
might be worth trying one of them at some point (see the "MySQL notes"
discussion of MySQL DB API drivers in the Django documentation).

svn path=/branches/tk705/; revision=6149