Whack rcynicng to produce an XML file which rcynic-html can parse: needs further work, but produces somewhat-useful output now.
Start merging command line argument (argparse) and config file
(ConfigParser) support into a unified library module, in an attempt to
reduce the amount of code duplication and reign in some of the
configuration chaos.  Trigger for this is all of the RP programs now
needing SQL configuration.

Whack rpki.rcynicdb.models to work with PostgreSQL, which has length
restrictions on indexed columns.

svn path=/branches/tk705/; revision=6239