rpkic needs to do uid swapping around XML reads too.
rpki.irdb.zookeeper.etree_read(), the functions that call it, and the
functions that call them, could all use a bit of refactoring.  At this
point pretty much ever caller is jumping through some hoop or another
due to the way the code has evolved, and most of it could be simpler.

svn path=/branches/tk705/; revision=6338