This time for sure.
We were building the correct management object structure in rpkidb,
but not in irdb, so the irdb kept breaking things when it tried to
bring them into alignment.  What we get for trying to perform an
upgrade which requires pushing on a rope.

Anyway, result now appears to work.  Committing, but letting it run
for a while before handing off to the usual alpha testers.

svn path=/branches/tk705/; revision=6422