New script to install an RPKI root key salvaged from an old rootd installation on top of the active CADetail of a new-style rootd-less rpkid configuration.
This has been tested, but only lightly, and only in the lab.

This script is dangerous.  Do not run it unless you really know what
you are doing, and even then you probably do not want to run it on
anything but a brand new installation with no existing RPKI data.

See #816.

svn path=/branches/tk705/; revision=6438