Accumulate primary keys instead of objects to work around Django caching.
This is nasty, and I still don't entirely understand it why this was
happening.  We collect ca_detail objects during bulk ROA processing,
so that we can defer manifest and CRL updates until the end of the
batch.  Somehow, Django's caching code was causing the parent CA's
issued serial number to roll back as part of this caching, which
caused us to reuse serial numbers.  Which is (very) bad.

Replacing the collection of ca_detail objects with a collection of
primary key values for those same ca_detail objects seems to have
worked, presumably because it lets us force creation of a new queryset
when it's time for us to process the relevant ca_detail objects.

The question is how many other booby traps like this might be lurking.