[max-length] Fix ROA parsing issue
The ROA update and generate methods were parsing the ipv4 and ipv6
instance variables as if they were resource sets, rather than ROA
prefix sets.  This works as expected when max-length is not being
used, but when max-length is used, parsing does not succeed and ROAs
cannot be generated or updated.  Parsing each variable as if it's a
ROA prefix set, and then converting the result to a resource set,
appears to fix the problem.